+++ src/main/java/org/yanbwe/raritycore/mixin/GuiGraphicsMixin.java
@@ -17,7 +17,8 @@ public abstract class GuiGraphicsMixin {
     @Inject(method = "renderItemDecorations(Lnet/minecraft/client/gui/Font;Lnet/minecraft/world/item/ItemStack;IILjava/lang/String;)V",
             at = @At(value = "TAIL"))
     private void renderItemDecorationsWithRarityBorder(Font font, ItemStack itemStack, int x, int y, @Nullable String s, CallbackInfo ci) {
-        if (!itemStack.isEmpty() && !RarityExclusionManager.isRenderingTooltipItem()) {
+        if (!itemStack.isEmpty() && !RarityExclusionManager.isRenderingTooltipItem()
+                && !RarityExclusionManager.isSuppressBorderRender()) {
             GuiGraphics guiGraphics = (GuiGraphics)(Object)this;
             ItemBorderRenderer.renderRarityBorder(guiGraphics, itemStack, x, y);
         }
@@ -26,7 +27,8 @@ public abstract class GuiGraphicsMixin {
     @Inject(method = "renderFakeItem(Lnet/minecraft/world/item/ItemStack;II)V",
             at = @At(value = "TAIL"))
     private void renderFakeItemWithRarityBorder(ItemStack itemStack, int x, int y, CallbackInfo ci) {
-        if (!itemStack.isEmpty() && !RarityExclusionManager.isRenderingTooltipItem()) {
+        if (!itemStack.isEmpty() && !RarityExclusionManager.isRenderingTooltipItem()
+                && !RarityExclusionManager.isSuppressBorderRender()) {
             GuiGraphics guiGraphics = (GuiGraphics)(Object)this;
             ItemBorderRenderer.renderRarityBorder(guiGraphics, itemStack, x, y);
         }